Helper: clean up linter warnings (actionable ones; skip the noisy linters)
Default `go vet` + `staticcheck` + default `golangci-lint` were
already mostly clean; running with --enable-all surfaced a mix of
real issues and stylistic noise. Fixed the real ones; intentionally
left the noisy linters' false positives.

## Fixed

main.go:112 (errcheck)
  flag.CommandLine.Parse(os.Args[1:]) had unchecked error. The
  default ExitOnError means a parse failure already exits, so the
  return is unreachable in practice — captured + logger.Fatalf
  so the linter and the eventual reader both see the intent.

compile.go:298 (gosimple S1008)
  `if len(p) > 500 { return false }; return true` → `return len(p) <= 500`.

pair.go:78 (errchkjson)
  json.Marshal of a struct of plain string/int64 cannot fail in
  practice but the linter wants it acknowledged. Acknowledged
  with an early-return on the unreachable error path.

llm.go:50 + server.go:114 + llm_test.go {/api/tags, /api/generate,
slow stream handler} (revive unused-parameter)
  Renamed unused http.Request / chunk-callback parameters to `_`.

trayicon_tray.go (gosec G115)
  int → uint32 conversions of `len(pixels)` etc. were flagged
  because gosec can't prove the values are statically bounded.
  Hoisted all the sizes into typed `const` declarations
  (pixelBytes, andMaskBytes, dataSize, entryOffset) computed
  from the hardcoded 16x16 dimensions. binary.Write now gets
  correctly-typed values; no conversions to flag.

trayicon_test.go (gosec G115)
  int32(binary.LittleEndian.Uint32(...)) compared to a positive
  literal — swapped for direct uint32 comparison (bit pattern
  matches for positive values). No conversion, no warning.

## Skipped — false positives or opinionated noise

goconst — wants every string with ≥3 occurrences turned into a
  constant ("127.0.0.1:9876", "darwin", "https", "dev"). Inline
  literals are more readable than `const TestHostPort = "..."`
  for these.

gosec G204 (exec.CommandContext on tex.Latexmk) — flagged because
  the binary path is a variable. The variable is the result of
  exec.LookPath in detectTex(), and args[] are hardcoded /
  validated. Already as safe as exec.* can be without removing
  the call entirely.

contextcheck (srv.HTTP.Shutdown(shutdownCtx)) — the warning
  recommends inheriting the parent context; we deliberately make
  a fresh WithTimeout context because the PARENT context was
  what triggered the shutdown signal and is already cancelled.
  This is the textbook graceful-shutdown pattern.

dupword on help text — flagged "rotate rotate" inside the help
  output, which is just "<subcommand>   <verb starting with
  that subcommand>". Correct natural prose.

errchkjson on json.Encode(w) for http.ResponseWriter — the
  response is best-effort; if the client disconnects mid-write
  there's nothing to do. Errors are intentionally discarded
  via `_ =`.

forbidigo — custom rule banning fmt.Print*. Standard Go; not part
  of the default linter set.

All helper tests pass (5.5s); cross-compiled darwin / linux /
windows clean.

+++ helper/trayicon_test.go
@@ -56,11 +56,14 @@ func TestTrayIconBytes_IsValidIcoStructure(t *testing.T) {
 	if hdr := binary.LittleEndian.Uint32(b[22:26]); hdr != 40 {
 		t.Errorf("BITMAPINFOHEADER.size = %d, want 40", hdr)
 	}
-	if w := int32(binary.LittleEndian.Uint32(b[26:30])); w != 16 {
+	// width / height are stored as int32 LE in the header. Compare
+	// against uint32 literals to skip the int32 cast — width and
+	// height are both positive in our case so the bit-pattern matches.
+	if w := binary.LittleEndian.Uint32(b[26:30]); w != 16 {
 		t.Errorf("BITMAPINFOHEADER.width = %d, want 16", w)
 	}
 	// height = h*2 = 32
-	if h := int32(binary.LittleEndian.Uint32(b[30:34])); h != 32 {
+	if h := binary.LittleEndian.Uint32(b[30:34]); h != 32 {
 		t.Errorf("BITMAPINFOHEADER.height = %d, want 32 (doubled for AND mask)", h)
 	}
